LPS-175848 Remove API called from Reduce and move to onClick function
