+++ src/manage-app-state/actionTypes.ts
@@ -1,6 +1,5 @@
 export enum TYPES {
 	SUBMIT_APP = 'SUBMIT_APP',
-	SUBMIT_APP_BUILD = 'SUBMIT_APP_BUILD',
 	SUBMIT_APP_LICENSING = 'SUBMIT_APP_LICENSING',
 	SUBMIT_APP_PROFILE = 'SUBMIT_APP_PROFILE',
 	SUBMIT_APP_STOREFRONT = 'SUBMIT_APP_STOREFRONT',
